rename: domain, service, dto 폴더 분리
